refactor: helper trait
