+++ crates/consensus/src/transaction/rlp.rs
@@ -1,0 +1,179 @@
+use crate::{SignableTransaction, Signed};
+use alloy_eips::eip2718::{Eip2718Error, Eip2718Result};
+use alloy_primitives::{keccak256, Parity, Signature};
+use alloy_rlp::{Buf, BufMut, Header};
+
+/// Helper trait for managing RLP encoding of transactions inside 2718
+/// envelopes.
+#[doc(hidden)]
+pub trait RlpEcdsaTx: SignableTransaction<Signature> + Sized {
+    /// The default transaction type for this transaction.
+    const DEFAULT_TX_TYPE: u8;
+
+    /// Calculate the encoded length of the transaction's fields, without a RLP
+    /// header.
+    fn rlp_encoded_fields_length(&self) -> usize;
+
+    /// Encodes only the transaction's fields into the desired buffer, without
+    /// a RLP header.
+    fn rlp_encode_fields(&self, out: &mut dyn alloy_rlp::BufMut);
+
+    /// Get the length of the transaction when RLP encoded.
+    fn rlp_encoded_length(&self) -> usize {
+        let payload_length = self.rlp_encoded_fields_length();
+        Header { list: true, payload_length }.length() + payload_length
+    }
+
+    /// RLP encodes the transaction.
+    fn rlp_encode(&self, out: &mut dyn BufMut) {
+        Header { list: true, payload_length: self.rlp_encoded_fields_length() }.encode(out);
+        self.rlp_encode_fields(out);
+    }
+
+    /// Get the length of the transaction when RLP encoded with the given
+    /// signature.
+    fn rlp_encoded_length_with_signature(&self, signature: &Signature) -> usize {
+        let payload_length = self.rlp_encoded_fields_length() + signature.rlp_vrs_len();
+        Header { list: true, payload_length }.length() + payload_length
+    }
+
+    /// RLP encodes the transaction with the given signature.
+    fn rlp_encode_signed(&self, signature: &Signature, out: &mut dyn BufMut) {
+        let payload_length = self.rlp_encoded_fields_length() + signature.rlp_vrs_len();
+        Header { list: true, payload_length }.encode(out);
+        self.rlp_encode_fields(out);
+        signature.write_rlp_vrs(out);
+    }
+
+    /// Get the length of the transaction when EIP-2718 encoded. This is the
+    /// 1 byte type flag + the length of the RLP encoded transaction.
+    fn eip2718_encoded_length(&self, signature: &Signature) -> usize {
+        self.rlp_encoded_length_with_signature(signature) + 1
+    }
+
+    /// EIP-2718 encode the transaction with the given signature and type flag.
+    fn eip2718_encode_with_type(&self, signature: &Signature, ty: u8, out: &mut dyn BufMut) {
+        out.put_u8(ty);
+        self.rlp_encode_signed(signature, out);
+    }
+
+    /// EIP-2718 encode the transaction with the given signature and the default
+    /// type flag.
+    fn eip2718_encode(&self, signature: &Signature, out: &mut dyn BufMut) {
+        self.eip2718_encode_with_type(signature, Self::DEFAULT_TX_TYPE, out);
+    }
+
+    /// Get the length of the transaction when network encoded. This is the
+    /// EIP-2718 encoded length with an outer RLP header.
+    fn network_encoded_length(&self, signature: &Signature) -> usize {
+        let payload_length = self.eip2718_encoded_length(signature);
+        Header { list: false, payload_length }.length() + payload_length
+    }
+
+    /// Network encode the transaction with the given signature.
+    fn network_encode_with_type(&self, signature: &Signature, ty: u8, out: &mut dyn BufMut) {
+        let payload_length = self.eip2718_encoded_length(signature);
+        Header { list: false, payload_length }.encode(out);
+        self.eip2718_encode_with_type(signature, ty, out);
+    }
+
+    /// Network encode the transaction with the given signature and the default
+    /// type flag.
+    fn network_encode(&self, signature: &Signature, out: &mut dyn BufMut) {
+        self.network_encode_with_type(signature, Self::DEFAULT_TX_TYPE, out);
+    }
+
+    /// Decodes the fields of the transaction from RLP bytes. Do not decode a
+    /// header. You may assume the buffer is long enough to contain the
+    /// transaction.
+    fn rlp_decode_fields(buf: &mut &[u8]) -> alloy_rlp::Result<Self>;
+
+    /// Decodes the transaction from RLP bytes.
+    fn rlp_decode(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {
+        let header = Header::decode(buf)?;
+        if !header.list {
+            return Err(alloy_rlp::Error::UnexpectedString);
+        }
+        let remaining_len = buf.len();
+
+        if header.payload_length > remaining_len {
+            return Err(alloy_rlp::Error::InputTooShort);
+        }
+
+        Self::rlp_decode_fields(buf)
+    }
+
+    /// Decodes the transaction from RLP bytes, including the signature.
+    fn rlp_decode_with_signature(buf: &mut &[u8]) -> alloy_rlp::Result<(Self, Signature)> {
+        let header = Header::decode(buf)?;
+        if !header.list {
+            return Err(alloy_rlp::Error::UnexpectedString);
+        }
+
+        let remaining = buf.len();
+        let tx = Self::rlp_decode_fields(buf)?;
+        let signature = Signature::decode_rlp_vrs(buf)?;
+
+        if !matches!(signature.v(), Parity::Parity(_)) {
+            return Err(alloy_rlp::Error::Custom("invalid parity for typed transaction"));
+        }
+
+        if buf.len() + header.payload_length != remaining {
+            return Err(alloy_rlp::Error::ListLengthMismatch {
+                expected: header.payload_length,
+                got: remaining - buf.len(),
+            });
+        }
+
+        Ok((tx, signature))
+    }
+
+    /// Decodes the transaction from RLP bytes, including the signature
+    /// Produces a [`Signed`].
+    fn rlp_decode_signed(buf: &mut &[u8]) -> alloy_rlp::Result<Signed<Self>> {
+        Self::rlp_decode_with_signature(buf).map(|(tx, signature)| tx.into_signed(signature))
+    }
+
+    /// Decodes the transaction from eip2718 bytes, expecting the given type
+    /// flag.
+    fn eip2718_decode_with_type(buf: &mut &[u8], ty: u8) -> Eip2718Result<Signed<Self>> {
+        let original_buf = *buf;
+
+        if buf.remaining() < 1 {
+            return Err(alloy_rlp::Error::InputTooShort.into());
+        }
+        let actual = buf.get_u8();
+        if actual != ty {
+            return Err(Eip2718Error::UnexpectedType(actual));
+        }
+
+        // OPT: We avoid re-serializing by calculating the hash directly
+        // from the original buffer contents.
+        let (tx, signature) = Self::rlp_decode_with_signature(buf)?;
+        let total_len = tx.eip2718_encoded_length(&signature);
+        let hash = keccak256(&original_buf[..total_len]);
+
+        Ok(Signed::new_unchecked(tx, signature, hash))
+    }
+
+    /// Decodes the transaction from eip2718 bytes, expecting the default type
+    /// flag.
+    fn eip2718_decode(buf: &mut &[u8]) -> Eip2718Result<Signed<Self>> {
+        Self::eip2718_decode_with_type(buf, Self::DEFAULT_TX_TYPE)
+    }
+
+    /// Decodes the transaction from network bytes.
+    fn network_decode_with_type(buf: &mut &[u8], ty: u8) -> Eip2718Result<Signed<Self>> {
+        let header = Header::decode(buf)?;
+        if header.list {
+            return Err(alloy_rlp::Error::UnexpectedList.into());
+        }
+        Self::eip2718_decode_with_type(buf, ty)
+    }
+
+    /// Decodes the transaction from network bytes, expecting the default type
+    /// flag.
+    fn network_decode(buf: &mut &[u8]) -> Eip2718Result<Signed<Self>> {
+        Self::network_decode_with_type(buf, Self::DEFAULT_TX_TYPE)
+    }
+}
